Split scanner.{c,h} into scanner.{c,h} and io.{c,h}
* scanner.{h,c} should now only contain code that is part of the
  GAP source scanner
* io.{h,c} should now contain code that deals with input that goes
  into the GAP interpreter, and output from the interpreter.

This requires some manoeuvering (and this commit has to be quite big
to keep things compiling). More cleanup and refactoring commits to
follow.
