+++ Makefile.rules
@@ -56,6 +56,7 @@ SOURCES += src/hookintrprtr.c
 SOURCES += src/integer.c
 SOURCES += src/intfuncs.c
 SOURCES += src/intrprtr.c
+SOURCES += src/io.c
 SOURCES += src/iostream.c
 SOURCES += src/listfunc.c
 SOURCES += src/listoper.c
