+++ src/lists.h
@@ -18,6 +18,7 @@
 #ifndef GAP_LISTS_H
 #define GAP_LISTS_H
 
+#include <src/io.h>      /* for Pr */
 #include <src/plist.h>
 
 extern  Obj             TYPE_LIST_EMPTY_MUTABLE;
